requirements.txt
jupyter==1.0.0
pandas~=1.3.5
google-cloud-storage==2.2.1
SQLAlchemy==1.3.0
psycopg2
pg8000==1.16.5
google-cloud-secret-manager